More updates based on PR feedback.
